requirements.txt
Django
djangorestframework
requests
django-environ 
python-decouple
django-redis
python-decouple
retrying
gunicorn